refactor(code): update file structure
