+++ src/components/Footer/Footer.module.scss
@@ -1,4 +1,4 @@
-@import '../../../variables.scss';
+@import '../../variables.scss';
 
 .footer {
   background-color: $footer_background;
