+++ src/components/Header/Header.module.scss
@@ -1,4 +1,4 @@
-@import '../../../variables.scss';
+@import '../../variables.scss';
 
 .header {
   display: flex;
