--- src/components/Layout/Layout.module.scss
@@ -1,1 +1,0 @@
-@import '../../variables.scss';
