+++ src/routes/Pages/Layout/index.ts
@@ -1,0 +1,3 @@
+import Layout from './Layout';
+
+export default Layout;
