+++ src/routes/Pages/NoMatch/index.ts
@@ -1,0 +1,3 @@
+import NoMatch from './NoMatch';
+
+export default { NoMatch };
