+++ src/routes/Pages/index.ts
@@ -2,5 +2,14 @@ import CharactersPage from './Characters';
 import ComicsPage from './Comics';
 import CharacterDetailsPage from './CharactersDetails';
 import ComicsDetailsPage from './ComicsDetails';
+import NoMatch from './NoMatch';
+import Layout from './Layout';
 
-export { CharactersPage, ComicsPage, CharacterDetailsPage, ComicsDetailsPage };
+export {
+  CharactersPage,
+  ComicsPage,
+  CharacterDetailsPage,
+  ComicsDetailsPage,
+  NoMatch,
+  Layout
+};
